requirements.txt
# Python requirements for ATIATO translation system
pyodbc
azure-identity
azure-keyvault-secrets
requests
